Fix parsing of detailed timings
According to VESA E-EDID 1.4 Section 3.10.3, only the first three bytes
of a display descriptor need to be zeros. The fifth byte is usually
zero, however, may also contain a value in case the descriptor tag
indicates that this is a Display Range Limits Descriptor.

Furthermore, the third byte may very well be zero even if the descriptor
contains a detailed timing: this is in fact the case for screens with,
e.g., 2560 or 3840 horizontal addressable pixels (i.e., WQHD or 4K UHD),
where the lower 8 bits of the horizontal addressable pixels are zeroes
(see Section 3.10.2).