EDID 1.4 model years
EDID 1.4 allows specifying a model year instead of week and year of
manufacturing. We kinda supported this behavior by ignoring the week
value when it was 0xff. However, we would still store and display the
model year as year of manufacturing. Considering that the model year
is a different piece of information, let's add a specific field for it.

Although EDID 1.3 didn't support model years, some vendors (Toshiba)
seem to use them anyway, so don't limit support to EDID 1.4.